refactor gui main window into modular panels
