+++ gui/widgets/__init__.py
@@ -1,0 +1,5 @@
+from .agent_panel import AgentPanel
+from .scan_panel import ScanPanel
+from .settings_panel import SettingsPanel
+
+__all__ = ["SettingsPanel", "AgentPanel", "ScanPanel"]
